+++ src/lib/services/spinner-visibility.service.ts
@@ -20,8 +20,13 @@ export class SpinnerVisibilityService {
     constructor(private pendingInterceptorService: PendingInterceptorService) {
     }
 
+    /** @deprecated Deprecated in favor of visibilityObservable$ */
     get visibilityObservable(): Observable<boolean> {
         return this._visibilitySubject.asObservable();
+    }
+
+    get visibilityObservable$(): Observable<boolean> {
+        return this.visibilityObservable;
     }
 
     public show(): void {