+++ src/test/services/spinner-visibility.service.spec.ts
@@ -23,12 +23,12 @@ describe('SpinnerVisibilityService', () => {
     }));
 
     it('should define a subject', inject([SpinnerVisibilityService], (service: SpinnerVisibilityService) => {
-        expect(service.visibilityObservable).toBeTruthy();
+        expect(service.visibilityObservable$).toBeTruthy();
     }));
 
     it('should pipe \'true\' when calling show()', inject([SpinnerVisibilityService], (spinner: SpinnerVisibilityService) => {
         spinner.show();
-        spinner.visibilityObservable.subscribe(result => {
+        spinner.visibilityObservable$.subscribe(result => {
                 expect(result).toBeTruthy();
             },
             error => {
@@ -38,7 +38,7 @@ describe('SpinnerVisibilityService', () => {
 
     it('should pipe \'false\' when calling hide()', inject([SpinnerVisibilityService], (spinner: SpinnerVisibilityService) => {
         spinner.hide();
-        spinner.visibilityObservable.subscribe(result => {
+        spinner.visibilityObservable$.subscribe(result => {
                 expect(result).toBeFalsy();
             },
             error => {